Notifications: 3 threads merged into 1
- all main methods of module moved in SchedulingManager
- NotificationListener deleted since all its functionality is moved to
  SchedulingManager
- tests altered with respect to the changes
- removed circular dependency of templateManager and regexManager
