+++ perun-notification/src/main/resources/perun-notification-applicationcontext.xml
@@ -27,12 +27,9 @@ http://www.springframework.org/schema/context http://www.springframework.org/sch
     <!-- 	<tx:annotation-driven transaction-manager="transactionManagerPerunController" /> -->
 
     <!-- Scans for @Repository, @Service and @Component -->
-    <context:component-scan base-package="cz.metacentrum.perun.notif.listener" />
     <context:component-scan base-package="cz.metacentrum.perun.notif.managers" />
     <context:component-scan base-package="cz.metacentrum.perun.notif.dao" />
 
-    <bean id="notificationListener" class="cz.metacentrum.perun.notif.listener.NotificationListener" autowire="byName">
-    </bean>
 
     <bean id="perunNotifAuditMessageDao" class="cz.metacentrum.perun.notif.dao.jdbc.PerunNotifAuditMessageDaoImpl" autowire="byName">
         <property name="dataSource" ref="dataSource"/>
